@@ -1,15 +1,10 @@
 import React, { useEffect } from 'react';
-import { useHistory, useLocation } from 'react-router';
+import { useNavigate } from 'react-router';
 import './BookmarkBar.scss';
 
 const BookmarkBar = () => {
-  const history = useHistory();
-  const location = useLocation();
-  useEffect(() => {
-    console.log(history);
-    console.log(location.state);
-  }, [history, location]);
-
+  const navigate = useNavigate();
+  
   return <div className="bookmark-bar">bookmark</div>;
 };
 
