@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, connect, useDispatch } from 'react-redux';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router';
+import { Button } from 'basicui';
 
 import './ChangeAsset.scss';
 
@@ -13,7 +14,7 @@ const ChangeAsset = (props: Props) => {
 
   const profile = useSelector((state: any) => state.profile);
 
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const assets = useSelector((state: any) => state.asset.assets);
 
@@ -32,16 +33,18 @@ const ChangeAsset = (props: Props) => {
   }, [props.space, assets]);
 
   const goToChangeAssetPage = () => {
-    history.push('/');
+    navigate('/');
   };
 
   return (
     <div className="change-asset">
       {currentAsset && <div>{currentAsset.name}</div>}
       <div>
-        <button onClick={goToChangeAssetPage}>
+        <Button
+          onClick={goToChangeAssetPage}
+        >
           {currentAsset ? 'Change company' : 'Choose company'}
-        </button>
+        </Button>
       </div>
     </div>
   );
