@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { useHistory } from 'react-router';
 import { addDays, format } from 'date-fns';
 import { faCheck, faPlus, faTimes } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -8,9 +7,7 @@ import './style.scss';
 import ReceiptModel from '../../../model/ReceiptModel';
 import ExpenseModel from '../../../model/ExpenseModel';
 import Topbar from '../../../components/Topbar';
-import DisableContextBarCommand from '../../../events/DisableContextBarCommand';
-
-const queryString = require('query-string');
+import { useNavigate } from 'react-router-dom';
 
 const EMPTY_EXPENSE: ExpenseModel = {
   amount: undefined,
@@ -34,22 +31,18 @@ interface Props {
 }
 
 const LandingPage = (props: Props) => {
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const authorization = useSelector((state: any) => state.authorization);
   const companyList = useSelector((state: any) => state.company.items);
 
   const goToCreateCompanyPage = () => {
-    history.push('/company/edit');
+    navigate('/company/edit');
   };
 
   const goToCompanyPage = (companyReference: number) => {
-    history.push(`/${companyReference}/note`);
+    navigate(`/${companyReference}/browse`);
   };
-
-  useEffect(() => {
-    DisableContextBarCommand.next(true);
-  }, []);
 
   return (
     <div className="landing-page">
