@@ -1,26 +1,26 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useHistory } from 'react-router';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCircleNodes, faListUl } from '@fortawesome/free-solid-svg-icons';
 import './style.scss';
 import Topbar from '../../../components/Topbar';
-import DisableContextBarCommand from '../../../events/DisableContextBarCommand';
+// import DisableContextBarCommand from '../../../events/DisableContextBarCommand';
 import { searchNote } from './service';
 import NoteModel from '../../../model/NoteModel';
 import SearchResults from '../../../components/SearchResults';
 import { isEmptyOrSpaces } from '../../../components/Utils';
 import GraphSearchResultsView from '../../../components/GraphSearchResultsView';
+import { useNavigate, useSearchParams } from 'react-router-dom';
+import { Input } from 'basicui';
+import MainSection from '../../../components/MainSection';
 
 interface Props {
   location: any;
   space: string;
 }
 
-const queryString = require('query-string');
-
 const SearchPage = (props: Props) => {
-  const history = useHistory();
+  const navigate = useNavigate();
   const profile = useSelector((state: any) => state.profile);
   const authorization = useSelector((state: any) => state.authorization);
   const dispatch = useDispatch();
@@ -30,42 +30,37 @@ const SearchPage = (props: Props) => {
   const [text, setText] = useState('');
   const [results, setResults] = useState<NoteModel[]>([]);
 
-  const [queryParam, setQueryParam] = useState<any>({});
+  const [searchParams] = useSearchParams();
+
+  // useEffect(() => {
+  //   DisableContextBarCommand.next(false);
+  // }, []);
 
   useEffect(() => {
-    DisableContextBarCommand.next(false);
-  }, []);
-
-  useEffect(() => {
-    const queryParam = queryString.parse(props.location.search);
-    setQueryParam(queryParam);
-  }, [props.location]);
-
-  useEffect(() => {
-    setText(queryParam?.text || '');
-    if (!isEmptyOrSpaces(queryParam?.text) && authorization.isAuth) {
-      searchNote(props.space, { text: queryParam.text }, authorization).then(
+    setText(searchParams.get('text') || '');
+    if (!isEmptyOrSpaces(searchParams.get('text')) && authorization.isAuth) {
+      searchNote(props.space, { text: searchParams.get('text') }, authorization).then(
         (response: NoteModel[]) => {
           setResults(response);
         }
       );
     }
-    if (isEmptyOrSpaces(queryParam?.text) && authorization.isAuth) {
+    if (isEmptyOrSpaces(searchParams.get('text')) && authorization.isAuth) {
       setResults([]);
     }
-  }, [queryParam, authorization]);
+  }, [searchParams, authorization]);
 
   const handleChange = (event: any) => {
     setText(event.target.value);
   };
 
   const handleChangeAndSubmit = (_text: string) => {
-    history.push(`/${props.space}/search?text=${_text}`);
+    navigate(`/${props.space}/search?text=${_text}`);
   };
 
   const handleSearch = (event: any) => {
     event.preventDefault();
-    history.push(`/${props.space}/search?text=${text}`);
+    navigate(`/${props.space}/search?text=${text}`);
     // saveNote(props.space, _note, authorization).then((response: any) => {
     //   if (response && id.current === response.reference) {
     //     setState(response);
@@ -101,17 +96,13 @@ const SearchPage = (props: Props) => {
           </button>
         </div>
       </Topbar>
-      <div className="search-page">
-        <form className="main" onSubmit={handleSearch}>
-          <input
-            className="text-lg search-page__input"
-            name="text"
+      <MainSection>
+        <form className="main search-page-form" onSubmit={handleSearch}>
+          <Input name="text"
             value={text}
             onInput={handleChange}
-            autoComplete="off"
-            placeholder="Type a text to search"
-            autoFocus
-          />
+            placeholder="Type to search"
+            autoFocus />
         </form>
         {view === 'list' && (
           <div className="search-page__results main">
@@ -127,7 +118,7 @@ const SearchPage = (props: Props) => {
             <GraphSearchResultsView space={props.space} noteNodes={results} />
           </div>
         )}
-      </div>
+      </MainSection>
     </div>
   );
 };
