@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router';
 import { format } from 'date-fns';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
@@ -10,15 +10,22 @@ import {
   faFileImport,
   faSkullCrossbones,
 } from '@fortawesome/free-solid-svg-icons';
+import { Input, Button, Checkbox } from 'basicui';
 import './BackupAndRestore.scss';
 import { newId } from '../../../events/MessageService';
 import * as service from './service';
 import RunLog from './RunLog';
 import { isEmptyAttributes } from '../../../components/Utils';
+import {
+  fetchAndAppendExpenseItems,
+  fetchAndSetExpenseItems,
+} from '../../../store/actions/ExpenseActions';
+import { fetchAllCategories } from '../../../store/actions/CategoryActions';
+import { fetchAllIncomeCategories } from '../../../store/actions/IncomeCategoryActions';
+import { fetchAllTags } from '../../../store/actions/TagActions';
 import Topbar from '../../../components/Topbar';
-import { fetchAndSetIncomeItems } from '../../../actions/IncomeActions';
-
-const queryString = require('query-string');
+import { fetchAndSetReceiptItems } from '../../../store/actions/ReceiptActions';
+import { fetchAndSetIncomeItems } from '../../../store/actions/IncomeActions';
 
 interface Props {
   space: string;
@@ -27,7 +34,7 @@ interface Props {
 
 const BackupAndRestore = (props: Props) => {
   const dispatch = useDispatch();
-  const history = useHistory();
+  const navigate = useNavigate();
   const authorization = useSelector((state: any) => state.authorization);
   const company = useSelector((state: any) =>
     state.company.items.find(
@@ -57,8 +64,8 @@ const BackupAndRestore = (props: Props) => {
     }
   }, [authorization]);
 
-  const handleChange = (detail: any) => {
-    setState({ ...state, [detail.name]: detail.value });
+  const handleFileChange = (event: any) => {
+    setState({ ...state, [event.target.name]: event.target.files });
   };
 
   const handleRunLogChange = (_data: any) => {
@@ -67,6 +74,28 @@ const BackupAndRestore = (props: Props) => {
   };
 
   const refreshStore = () => {
+    dispatch(
+      fetchAndSetExpenseItems(props.space, authorization, {
+        ...expenseFilter,
+        pagination: {
+          ...expensePagination,
+          pageSize: 20,
+          pageNo: 0,
+          hasMore: true,
+        },
+      })
+    );
+    dispatch(
+      fetchAndSetReceiptItems(props.space, authorization, {
+        ...receiptFilter,
+        pagination: {
+          ...receiptPagination,
+          pageSize: 20,
+          pageNo: 0,
+          hasMore: true,
+        },
+      })
+    );
     dispatch(
       fetchAndSetIncomeItems(props.space, authorization, {
         ...incomeFilter,
@@ -78,6 +107,9 @@ const BackupAndRestore = (props: Props) => {
         },
       })
     );
+    dispatch(fetchAllCategories(props.space, authorization));
+    dispatch(fetchAllIncomeCategories(props.space, authorization));
+    dispatch(fetchAllTags(props.space, authorization));
   };
 
   const importExpenseFile = () => {
@@ -133,27 +165,27 @@ const BackupAndRestore = (props: Props) => {
             <div className="page-title">Data export</div>
             <div>Export all expense data available in the system so far</div>
             <div className="backup-and-restore__section__action">
-              <button onClick={exportExpenseFile}>
+              <Button onClick={exportExpenseFile}>
                 <FontAwesomeIcon icon={faFileExport} />
                 Export
-              </button>
+              </Button>
             </div>
           </div>
           <div className="content-section">
             <div className="page-title">Data import</div>
             <div>
-              <input
+              <Input
                 name="expenseImportFile"
                 type="file"
                 value={state.file}
-                onChange={handleChange}
+                onInput={handleFileChange}
               />
               {state.expenseImportFile?.length > 0 && (
                 <div className="backup-and-restore__section__action">
-                  <button onClick={importExpenseFile}>
+                  <Button onClick={importExpenseFile}>
                     <FontAwesomeIcon icon={faFileImport} />
                     Import
-                  </button>
+                  </Button>
                 </div>
               )}
             </div>
@@ -174,10 +206,10 @@ const BackupAndRestore = (props: Props) => {
               of the data before performing this action.
             </div>
             <div className="backup-and-restore__section__action">
-              <button onClick={() => {}}>
+              <Button onClick={() => { }}>
                 <FontAwesomeIcon icon={faSkullCrossbones} />
                 Delete all data
-              </button>
+              </Button>
             </div>
           </div>
         </div>
