@@ -1,15 +1,14 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrash } from '@fortawesome/free-solid-svg-icons';
+import { Button, Input } from 'basicui';
 
 import './Permissions.scss';
 import { newId } from '../../../events/MessageService';
 import { getUserInvite, sendUserInvite } from './service';
 import Topbar from '../../../components/Topbar';
-
-const queryString = require('query-string');
 
 interface Props {
   space: string;
@@ -17,7 +16,7 @@ interface Props {
 }
 
 const Permissions = (props: Props) => {
-  const history = useHistory();
+  const navigate = useNavigate();
   const authorization = useSelector((state: any) => state.authorization);
   const users = useSelector((state: any) => state.user.items);
   const [userInviteList, setUserInviteList] = useState<any[]>([]);
@@ -41,8 +40,8 @@ const Permissions = (props: Props) => {
     }
   }, [authorization]);
 
-  const handleChange = (detail: any) => {
-    setState({ ...state, [detail.name]: detail.value });
+  const handleChange = (event: any) => {
+    setState({ ...state, [event.currentTarget.name]: event.currentTarget.value });
   };
 
   const addUser = (event: any) => {
@@ -64,21 +63,23 @@ const Permissions = (props: Props) => {
       <Topbar title="User administration" />
       <div className="main-section">
         <div className="permissions page-width content-section">
-          <form onSubmit={addUser}>
+          <form id={formId} onSubmit={addUser}>
             <div className="permissions__form">
-              <input
+              <Input
                 name="email"
                 value={state.email}
-                onChange={handleChange}
+                onInput={handleChange}
                 placeholder="Invite an user by e-mail"
-                autoFocus
+                autofocus
                 required
               />
-              <button onClick={addUser}>Invite user</button>
+              <Button type="submit" onClick={addUser}>
+                Invite user
+              </Button>
             </div>
           </form>
           <div className="permissions__list">
-            <table>
+            <table className="basicui-table">
               <thead>
                 <tr>
                   <th>First name</th>
