@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router';
 import { addDays, format } from 'date-fns';
 import {
   faBuilding,
@@ -12,6 +12,7 @@ import {
   faTimes,
 } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { useNavigate, useSearchParams } from 'react-router-dom';
 import './style.scss';
 import ReceiptModel from '../../../model/ReceiptModel';
 import ExpenseModel from '../../../model/ExpenseModel';
@@ -21,35 +22,28 @@ import EditCompany from './EditCompany';
 import Permissions from './Permissions';
 import BackupAndRestore from './BackupAndRestore';
 
-const queryString = require('query-string');
-
 interface Props {
   space: string;
   location: any;
 }
 
 const SettingsPage = (props: Props) => {
-  const history = useHistory();
+  const navigate = useNavigate();
   const authorization = useSelector((state: any) => state.authorization);
-  const [queryParam, setQueryParam] = useState<any>({});
-
-  useEffect(() => {
-    const query = queryString.parse(props.location.search);
-    setQueryParam({ ...query });
-  }, [props.location.search]);
+  const [searchParams, setSearchParams] = useSearchParams();
 
   return (
     <div className="settings-page">
       <Topbar title="Settings" />
       <div className="settings-page-container main-section">
         {/* <div className="settings-page__main"> */}
-        {queryParam.link === 'general' && (
+        {searchParams.get('link') === 'general' && (
           <EditCompany space={props.space} location={props.location} />
         )}
-        {queryParam.link === 'permissions' && (
+        {searchParams.get('link') === 'permissions' && (
           <Permissions space={props.space} location={props.location} />
         )}
-        {queryParam.link === 'backup' && (
+        {searchParams.get('link') === 'backup' && (
           <BackupAndRestore space={props.space} location={props.location} />
         )}
         {/* </div> */}
