@@ -1,26 +1,20 @@
-import 'react-app-polyfill/ie9';
-import 'react-app-polyfill/ie11';
-import 'react-app-polyfill/stable';
+// import 'react-app-polyfill/ie9';
+// import 'react-app-polyfill/ie11';
+// import 'react-app-polyfill/stable';
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-import '../node_modules/codemirror/addon/hint/show-hint';
-import '../node_modules/codemirror/addon/hint/anyword-hint';
-
-import './styles/index.css';
-// import { CookiesProvider } from 'react-cookie';
+import './basicui-styles/index.scss';
 import App from './components/App';
-import * as serviceWorker from './serviceWorker';
+// import './index.scss';
+// import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
-  //   <CookiesProvider>
   <App />,
-  //   </CookiesProvider>,
   document.getElementById('root')
 );
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+// serviceWorker.unregister();
