@@ -1,13 +1,23 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
+import ExpenseFilterModel from '../../model/ExpenseFilterModel';
+import {
+  ExpenseFilterPaginationState,
+  ExpenseFilterState,
+} from '../../simplestates/ExpenseFilterState';
+import ExpenseListState from '../../simplestates/ExpenseListState';
+import { fetchAllCategories } from '../../store/actions/CategoryActions';
 import { receiveMessage, sendMessage } from '../../events/MessageService';
-import { fetchAndSetCompanyItems } from '../../actions/CompanyActions';
-import { fetchAndSetUserItems } from '../../actions/UserActions';
-import { setProfile } from '../../actions/ProfileActions';
+import { searchExpense } from '../Page/ExpensePage/service';
+import PaginationModel from '../../model/PaginationModel';
+import ExpenseStateActions from '../../simplestates/ExpenseStateActions';
+import { fetchAndSetCompanyItems } from '../../store/actions/CompanyActions';
+import { fetchAndSetUserItems } from '../../store/actions/UserActions';
+import { setProfile } from '../../store/actions/ProfileActions';
+import ReceiptStateActions from '../../simplestates/ReceiptStateActions';
 import IncomeStateActions from '../../simplestates/IncomeStateActions';
-import { fetchAndSetFolderItems } from '../../actions/FolderActions';
-import { fetchAndSetNoteItems } from '../../actions/NoteActions';
-import { fetchAndSetTagItems } from '../../actions/TagActions';
+import { fetchAndSetNoteItems } from '../../store/actions/NoteActions';
+import { useParams } from 'react-router-dom';
 
 const Init = () => {
   const authorization = useSelector((state: any) => state.authorization);
@@ -22,10 +32,9 @@ const Init = () => {
     if (authorization?.isAuth && space) {
       //  && !previousAuthorizationState?.isAuth) {
       initialize();
+      initializeHttpInterceptor();
       dispatch(fetchAndSetUserItems(space, authorization));
-      dispatch(fetchAndSetFolderItems(space, authorization));
       dispatch(fetchAndSetNoteItems(space, authorization));
-      dispatch(fetchAndSetTagItems(space, authorization));
     }
   }, [authorization, space]);
 
@@ -45,32 +54,44 @@ const Init = () => {
   useEffect(() => {
     initializeProfileFromSession();
     receiveMessage().subscribe((event: any) => {
-      console.log(event);
       if (event.name === 'spaceChange') {
+        // TODO
         setSpace(event.data);
       }
       if (event.name === 'spaceChange' && authorization.isAuth) {
+        setSpace(event.data);
         initialize();
+        initializeHttpInterceptor();
       }
     });
   }, []);
 
   // useEffect(() => {
-  //   console.log('profile.theme');
-  //   if (profile.theme === 'theme_light') {
-  //     document.body.style.backgroundColor = 'var(--color-global-lightmode)';
-  //   } else {
-  //     document.body.style.backgroundColor = 'var(--color-global-darkmode)';
-  //   }
-  // }, [profile.theme]);
+  //   document.body.addEventListener('mousedown', () => {
+  //     sendMessage('usingMouse', true);
+  //   });
+
+  //   // Re-enable focus styling when Tab is pressed
+  //   document.body.addEventListener('keydown', (event: any) => {
+  //     if (event.keyCode === 9) {
+  //       sendMessage('usingMouse', false);
+  //     }
+  //   });
+  // }, [profile]);
 
   useEffect(() => {
-    if (profile.theme === 'theme_light') {
-      document.documentElement.classList.add('light');
-      document.documentElement.classList.remove('dark');
+    if (profile.theme === 'basicui-light') {
+      document.body.classList.add("basicui-light");
+      document.body.classList.add("writeup-light");
+      document.body.classList.remove("basicui-dark");
+      document.body.classList.remove("writeup-dark");
+      // document.body.style.backgroundColor = 'var(--theme-white-50)';
     } else {
-      document.documentElement.classList.add('dark');
-      document.documentElement.classList.remove('light');
+      document.body.classList.add("basicui-dark");
+      document.body.classList.add("writeup-dark");
+      document.body.classList.remove("basicui-light");
+      document.body.classList.remove("writeup-light");
+      // document.body.style.backgroundColor = 'var(--theme-black-800)';
     }
   }, [profile.theme]);
 
@@ -83,22 +104,75 @@ const Init = () => {
 
   const initializeProfileFromSession = () => {
     const colorMode = sessionStorage.getItem('neuralweb_pref_profile_colormode');
-    const sidebarStatus = sessionStorage.getItem(
-      'neuralweb_pref_sidebar_status'
-    );
+    const sidebarStatus = sessionStorage.getItem('neuralweb_pref_sidebar_status');
 
     if (colorMode || sidebarStatus) {
       dispatch(
         setProfile({
-          theme: colorMode || 'theme_dark',
+          theme: colorMode || 'basicui-dark basicui-dark',
           sidebar: sidebarStatus === 'expanded',
         })
       );
     }
   };
 
+  const initializeHttpInterceptor = () => {
+    console.log('HTTP Interceptor initialization');
+    // TODO
+    // axiosInstance.defaults.headers.authorization = authorization.access_token;
+    // axiosInstance.interceptors.response.use(
+    //   (response) => {
+    //     return response;
+    //   },
+    //   (error) => {
+    //     if (error.response.status !== 401) {
+    //       return new Promise((resolve, reject) => {
+    //         reject(error);
+    //       });
+    //     }
+    //     httpPost(
+    //       '/auth/token',
+    //       {
+    //         refresh_token: authorization.refresh_token,
+    //         grant_type: 'refresh_token',
+    //         realm: realm || 100,
+    //       },
+    //       null
+    //     )
+    //       .then((response) => {
+    //         if (response.status === 200) {
+    //           axiosInstance.defaults.headers.authorization =
+    //             response.data.access_token;
+    //           setSessionValue(
+    //             `neuralweb-access_token`,
+    //             response.data.access_token
+    //           );
+    //           dispatch(
+    //             addAuth({
+    //               ...authorization,
+    //               access_token: response.data.access_token,
+    //             })
+    //           );
+    //           if (!error.config._retry) {
+    //             error.config._retry = true;
+    //             error.config.headers.authorization = response.data.access_token;
+    //             return axiosInstance(error.config);
+    //           }
+    //         } else {
+    //           console.log('********redirect to login');
+    //         }
+    //       })
+    //       .catch((error) => {
+    //         Promise.reject(error);
+    //       });
+    //   }
+    // );
+  };
+
   return (
     <>
+      <ExpenseStateActions space={space} />
+      <ReceiptStateActions space={space} />
       <IncomeStateActions space={space} />
     </>
   );
