@@ -1,10 +1,12 @@
 import axios from 'axios';
 
+export const axiosInstance = axios.create();
+export const axiosManualInstance = axios.create();
+
 const baseUrl = process.env.REACT_APP_API_URL;
 
 export function httpGet(endpoint: string, headers: any, url?: string) {
-  console.log('***************');
-  return axios.get((url || baseUrl) + endpoint, headers);
+  return axiosInstance.get((url || baseUrl) + endpoint, headers);
   // .then(function(response) {
   //     return Promise.resolve(response);
   // }
@@ -17,7 +19,7 @@ export function httpPost(
   headers: any,
   url?: string
 ) {
-  return axios.post((url || baseUrl) + endpoint, payload, headers);
+  return axiosInstance.post((url || baseUrl) + endpoint, payload, headers);
   //     .then(function(response) {
   //         return Promise.resolve(response);
   //     }
@@ -30,15 +32,78 @@ export function httpPut(
   headers: any,
   url?: string
 ) {
-  return axios.put((url || baseUrl) + endpoint, payload, headers);
+  return axiosInstance.put((url || baseUrl) + endpoint, payload, headers);
   //     .then(function(response) {
   //         return Promise.resolve(response);
   //     }
   // )
 }
 
-export function httpDelete(endpoint: string, headers: any, url?: string) {
-  return axios.delete((url || baseUrl) + endpoint, headers);
+export function httpDelete(
+  endpoint: string,
+  headers: any,
+  url?: string
+) {
+  return axiosInstance.delete((url || baseUrl) + endpoint, headers);
+  // return axiosInstance.delete((url || baseUrl) + endpoint, {
+  //   headers,
+  //   data: payload,
+  // });
+  //     .then(function(response) {
+  //         return Promise.resolve(response);
+  //     }
+  // )
+}
+
+export function httpGetManual(endpoint: string, headers: any, url?: string) {
+  return axiosManualInstance.get((url || baseUrl) + endpoint, headers);
+  // .then(function(response) {
+  //     return Promise.resolve(response);
+  // }
+  // )
+}
+
+export function httpPostManual(
+  endpoint: string,
+  payload: any,
+  headers: any,
+  url?: string
+) {
+  return axiosManualInstance.post(
+    (url || baseUrl) + endpoint,
+    payload,
+    headers
+  );
+  //     .then(function(response) {
+  //         return Promise.resolve(response);
+  //     }
+  // )
+}
+
+export function httpPutManual(
+  endpoint: string,
+  payload: any,
+  headers: any,
+  url?: string
+) {
+  return axiosManualInstance.put((url || baseUrl) + endpoint, payload, headers);
+  //     .then(function(response) {
+  //         return Promise.resolve(response);
+  //     }
+  // )
+}
+
+export function httpDeleteManual(
+  endpoint: string,
+  payload: any,
+  headers: any,
+  url?: string
+) {
+  return axiosManualInstance.delete((url || baseUrl) + endpoint);
+  // return axiosInstance.delete((url || baseUrl) + endpoint, {
+  //   headers,
+  //   data: payload,
+  // });
   //     .then(function(response) {
   //         return Promise.resolve(response);
   //     }
