@@ -7,12 +7,8 @@ import Notification from '../Notification';
 import NavigationContainer from '../App/NavigationContainer';
 import BodyContainer from '../App/BodyContainer';
 import SideContent from './SideContent';
-import FileExplorer from '../FileExplorer';
-import SideContentMini from './SideContentMini';
-import DisableContextBarCommand from '../../events/DisableContextBarCommand';
 
 interface Props {
-  cookies: any;
   space: string;
 }
 
@@ -21,30 +17,21 @@ const MainContent = (props: Props) => {
   const authorization = useSelector((state: any) => state.authorization);
   const dispatch = useDispatch();
 
-  const [disableContextBar, setDisableContextBar] = useState(false);
-
-  useEffect(() => {
-    DisableContextBarCommand.asObservable().subscribe((message) => {
-      setDisableContextBar(message);
-    });
-  }, []);
-
   return (
     <>
-      <SideContentMini cookies={props.cookies} space={props.space} />
-      <SideContent cookies={props.cookies} space={props.space} />
-      {/* <FileExplorer space={props.space} /> */}
+      <div className="desktop-only">
+        <SideContent space={props.space} />
+      </div>
       {/* <NavigationContainer
         cookies={props.cookies}
         space={props.space}
         transparent={false}
       /> */}
       <div
-        className={`main-content ${
-          profile.sidebar
+        className={`main-content ${profile.sidebar
             ? 'main-content__sidebar-active'
             : 'main-content__sidebar-inactive'
-        }`}
+          }`}
       >
         <BodyContainer {...props} />
       </div>
