@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import './DarkModeIcon.scss';
 import { NightsStay, WbSunny } from '@material-ui/icons';
-import { setProfile } from '../../actions/ProfileActions';
+import { setProfile } from '../../store/actions/ProfileActions';
 
 const DarkModeIcon = () => {
   const profile = useSelector((state: any) => state.profile);
@@ -10,19 +10,19 @@ const DarkModeIcon = () => {
   const toggleMode = () => {
     dispatch(
       setProfile({
-        theme: profile.theme === 'theme_dark' ? 'theme_light' : 'theme_dark',
+        theme: profile.theme === 'basicui-dark' ? 'basicui-light' : 'basicui-dark',
       })
     );
 
     sessionStorage.setItem(
       'neuralweb_pref_profile_colormode',
-      profile.theme === 'theme_dark' ? 'theme_light' : 'theme_dark'
+      profile.theme === 'basicui-dark' ? 'basicui-light' : 'basicui-dark'
     );
   };
   return (
     <div className={`dark-mode-icon ${profile.theme}`} onClick={toggleMode}>
-      {profile.theme === 'theme_dark' && <WbSunny className="cursor-pointer" />}
-      {profile.theme !== 'theme_dark' && (
+      {profile.theme === 'basicui-dark' && <WbSunny className="cursor-pointer" />}
+      {profile.theme !== 'basicui-dark' && (
         <NightsStay className="cursor-pointer" />
       )}
     </div>
