@@ -2,19 +2,27 @@ import { faBars } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { setProfile } from '../../actions/ProfileActions';
+import { setProfile } from '../../store/actions/ProfileActions';
+import MobileSidebar from '../MobileSidebar';
 import './style.scss';
 
 interface Props {
   title: string;
   children?: any;
-  fixed?: boolean;
-  isContextExpanded?: boolean;
 }
 
 const Topbar = (props: Props) => {
   const dispatch = useDispatch();
   const profile = useSelector((state: any) => state.profile);
+  const [isMobileSidebarOpen, setIsMobileSidebarOpen] = useState(false);
+
+  const toggleMobileSidebar = () => {
+    setIsMobileSidebarOpen(!isMobileSidebarOpen);
+  }
+
+  const closeMobileSidebar = () => {
+    setIsMobileSidebarOpen(false);
+  }
 
   const toggleSidebar = () => {
     sessionStorage.setItem(
@@ -26,22 +34,21 @@ const Topbar = (props: Props) => {
   };
 
   return (
-    <div
-      className={`topbar bg-light-300 text-gray-900 dark:bg-dark-400 dark:text-gray-200 ${
-        props.fixed ? 'topbar--fixed' : 'topbar--not-fixed'
-      } ${
-        props.isContextExpanded
-          ? 'topbar--context-active'
-          : 'topbar--context-inactive'
-      } ${
-        profile.sidebar ? 'topbar--sidebar-active' : 'topbar--sidebar-inactive'
-      }`}
-    >
-      <div className="topbar__left">
-        <div>{props.title}</div>
+    <>
+      <div className="topbar">
+        <div className="topbar__left">
+          <button className="button desktop-only" onClick={toggleSidebar}>
+            <FontAwesomeIcon icon={faBars} />
+          </button>
+          <button className="button mobile-only" onClick={toggleMobileSidebar}>
+            <FontAwesomeIcon icon={faBars} />
+          </button>
+          <div>{props.title}</div>
+        </div>
+        <div className="topbar__right">{props.children}</div>
       </div>
-      <div className="topbar__right">{props.children}</div>
-    </div>
+      <MobileSidebar isOpen={isMobileSidebarOpen} onClose={closeMobileSidebar} />
+    </>
   );
 };
 
